+++ internal/repo/chain/repo.go
@@ -3,8 +3,8 @@ package chain
 import "context"
 
 type Repo interface {
-	Push(ctx context.Context, userID int64, state string) error
-	Pop(ctx context.Context, userID int64) (string, error)
-	LastState(ctx context.Context, userID int64) (string, error)
-	Clear(ctx context.Context, userID int64) error
+	Push(ctx context.Context, userID int, state string) error
+	Pop(ctx context.Context, userID int) (string, error)
+	LastState(ctx context.Context, userID int) (string, error)
+	Clear(ctx context.Context, userID int) error
 }
